Modernize; move entirely over to cmake.

+++ board.c
@@ -12,9 +12,8 @@ char board[HEIGHT][WIDTH] = {{'P', 'E', 'R', 'S'},
 /** Shuffle order of dice */
 
 static void shuffle(const char *array[], int n) {
-    for (int i = 0; i < n - 1; i++) {
-	// NOLINT(cppcoreguidelines-narrowing-conversions)
-        int j = i + random() % (n - i); 
+    for (long i = 0; i < n - 1; i++) {
+        long j = i + random() % (n - i);
         const char *temp = array[j];
         array[j] = array[i];
         array[i] = temp;
@@ -25,9 +24,9 @@ static void shuffle(const char *array[], int n) {
 
 void make_board() {
     const char *dice[] = {
-            "AAEEGN", "ELRTTY", "AOOTTW", "ABBJOO", "EHRTVW", "CIMOTU", 
-	    "DISTTY", "EIOSST", "DELRVY", "ACHOPS", "HIMNQU", "EEINSU", 
-	    "EEGHNW", "AFFKPS", "HLNNRZ", "DEILRX"};
+            "AAEEGN", "ELRTTY", "AOOTTW", "ABBJOO", "EHRTVW", "CIMOTU",
+            "DISTTY", "EIOSST", "DELRVY", "ACHOPS", "HIMNQU", "EEINSU",
+            "EEGHNW", "AFFKPS", "HLNNRZ", "DEILRX"};
     srandom((unsigned int) time(NULL));
     shuffle(dice, 16);
 
@@ -35,4 +34,3 @@ void make_board() {
         for (int x = 0; x < WIDTH; x++)
             board[y][x] = dice[y * 4 + x][random() % 6];
 }
-
